feat(core): lift Parquet writer into cqlite-core behind a parquet feature (#683, #684, #685)
Move the batch and streaming Parquet writers from cqlite-cli into
cqlite-core/src/export/parquet.rs behind a new optional 'parquet' cargo
feature (off by default; requires state_machine):

- Decouple from CLI-only types (#683): writer-owned ParquetExportOptions
  (row limit, row group size, compression) replaces OutputConfig; a
  dedicated thiserror ParquetExportError replaces OutputError in writer
  signatures; ValueFormatter moves to cqlite_core::util::value_fmt (the
  CLI re-exports it unchanged); no unwrap()/expect() outside tests.
- Core feature + move (#684): arrow/parquet become optional cqlite-core
  dependencies wired to the new feature; the default build does not
  compile them (verified via cargo tree). Module docs state the feature
  flag and the external-committer boundary (CQLite produces Parquet
  files; Iceberg/Delta commits are out of scope). The writer's unit
  tests move alongside the code.
- CLI re-wire (#685): cqlite-cli/src/output/parquet.rs is now a thin
  adapter preserving the historical API (ParquetWriter::write taking
  OutputConfig, create_streaming_parquet_writer taking a row group
  size, StreamingWriter trait impl) and mapping core errors to
  OutputError exactly as before. arrow/parquet move to dev-dependencies
  (read-back validation only). A new byte-for-byte golden test
  (tests/parquet_golden_tests.rs, regenerable via
  UPDATE_PARQUET_GOLDEN=1) guards the output; all 111 pre-existing
  parquet writer tests and the 12 dataset round-trip tests pass
  unmodified.

Part of Epic #682.
