+++ cqlite-core/src/util/mod.rs
@@ -1,3 +1,4 @@
 //! Utility modules shared across cqlite-core
 
 pub mod cassandra_murmur3;
+pub mod value_fmt;
